refactor: 불필요하게 분리한 View 제거를 통한 ViewController 기능 복원
